JuliaDB -> DataFrames
Read the database with CSV and provide a DataFrame view.
This allows using the familiar DataFrame syntax to sift through the
dataset. However, CSV doesn't appear to consider the same rows of the
database as valid, so some matrices that used to be available are no
longer available (e.g., the Arenas matrices). Thus, this is a breaking
update.

Main differences:
* the database is read with CSV and a DataFrame "view" is provided;
* a selection of matrices always returns a DataFrame or a DataFrameRow;
* group_path and matrix_path become group_paths and matrix_paths because
  those methods always return an array of paths.